Failure

+++ srcs/FdF.h
@@ -6,7 +6,7 @@
 /*   By: ahernand <ahernand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/16 18:33:13 by ahernand          #+#    #+#             */
-/*   Updated: 2022/10/18 12:18:13 by ahernand         ###   ########.fr       */
+/*   Updated: 2022/10/19 14:11:34 by ahernand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,8 +45,8 @@ typedef struct		t_dt
 
 	int				fullcreen;
 
-	int				highest_y;
-	int				lowest_y;
+	double			highest_y;
+	double			lowest_y;
 
 
 	double			line_height;
@@ -61,8 +61,8 @@ typedef struct		t_dt
 	int				**lines;
 
 
-	int				**i;
-	int				**j;
+	double			**i;
+	double			**j;
 }					t_dt;
 
 /*
